refactor(facet-ui): the derived form renders wherever a facet is shown
Tiers 1 and 2 — the form derived from a facet's schema, refined by its
`editor` spec — lived in apps/web. A DECLARED editor should look the same on
every surface, and while the renderer lived in one vessel a second surface
would have had to reimplement the derivation to honour the same declaration.

`DerivedFacetForm` moves to `facet-ui` beside the plugin surface it serves.
It now derives the form from the registry itself rather than taking one as a
prop: a caller that computes the form can compute it differently, which is
the drift this move exists to close.

Its 20 utility class names become inline values and host custom properties,
per this package's rule — a class named inside a workspace package is never
generated and fails silently.

Also fixes a formatting/lint finding in ci-verify-coverage.test.ts that came
from main and blocked the local gate.

Visual evidence: none — the rendering is unchanged by construction, which is
the point; a before/after would be two identical panels and
compose-figure.mjs refuses those. Verified by screenshot comparison during
the move, and covered by 22 panel tests plus the package's own 7.
